sweethome3d: fix impurity
The sweethome3d program ships pre-compiled shared libraries for graphics that
need to be `patchelf`ed to work properly. Once the libraries are compressed into
the final .jar, Nix can no longer see the path references so they are not
declared a runtime dependency and the program won't run on a system that didn't
build it and doesn't happen to have the right paths already present on it by
chance.

This patch switches to autoPatchelfHook to ensure the libraries get patched
correctly, then scans the libraries to make a plain-text file with the same
paths so that Nix can see the references and account for them. I suspect this
problem might be lurking in other Java programs in nixpkgs.

Fixes #66471.