nixos/utils: Let escapeSystemdPath be correct for more paths
I was working with the systemd services for `networking.wireguard.interfaces.<name>.peers.*`
and noticed that `utils.escapeSystemdPath` didn't escape + or =.

This commit expands the range of characters which `escapeSystemdPath`
can correctly escape. It also adds some type checking, plus handling
of a trailing slash, leading dot, and the special case for /.

`escapeSystemdPath` is now defined in terms of a new function
`escapeSystemd`.

Some unit tests have been added to the nixos tests.

I have removed the "FIXME: slow" comment because I believe that it is
less relevant now than previously. This is because `escapeSystemdPath`
was last touched in
3f6d53cc972fe8bf34b5cb0184087d95c1130c1d (2012-10-13). But since then,
`replaceChars` was modified to use the `replaceStrings` primop in
9cfd128a4219ba9d1d143852fdf25be0710f97f3 (2015-07-24).

Finally, I added some code comments and a `traceIf` warning for the
situations where `escapeSystemdPath` still doesn't return correct
results. I don't believe it's feasible to make a pure nix
`escapeSystemdPath` 100% correct, nor is it really necessary. But if
correctness is truly needed, then the best way is probably by adding a
new builtin function to nix.