nextcloud24: init at 24.0.0
Added Nextcloud 23 and set it as the default Nextcloud version for the
NixOS module. Added PHP 8.1 as an option for phpPackage and default for
Nextcloud ≥ 24.

(cherry picked from commit ecd8d42397d0142b9b54e1e0664acb43c146fe98)

Backport #171736 to release-21.11. We have multiple attributes for
different majors, so this is technically a new package that can be
backported.

The defaults aren't changed, so nextcloud22 is still the default for
NixOS 21.11. We have also removed php81 references, as that has not been
backported to 21.11.