ghcHEAD: Fix Windows cross-compilation with lld.
By specifying pkgs.libffi here instead of letting callPackage handle
it, we confuse the splicing logic and put the host libffi into
NIX_LDFLAGS_FOR_TARGET. This previously hasn't been a problem, as we
also pass an explicit configure flag pointing to the target libffi,
and so the only side-effect is a senseless -rpath flag.

ld.lld (rightly) does not recognize the -rpath flag when targeting
Windows, however, so this causes build failures.