ncurses: make static binaries not depend on Nix store paths
One of the most popular features of static binaries is portability
across distros. However, static binaries built against the ncurses
package are currently not portable. This is because the current ncurses
package configures dependant packages to use the ncurses store path as
their terminfo search path. This unfortunately makes static binaries
built against ncurses not portable across NixOS systems too, as it
introduces a dependency on a specific build of ncurses.

To fix this problem, this change adds common paths from FHS systems and
the NixOS system profile path to the default terminfo search path.