libmodsecurity: remove dependency on valgrind
valgrind is only supported on certain platforms (it notably lacks support for
armv6l), so needs to be excluded as a dependency on unsupported platforms. In
this case, it turns out that valgrind is not currently used at all, so I
removed it entirely. libmodsecurity can optionally use valgrind to run
tests, but we don't run the tests at all.