lib/types.nix: listOf.check needs to check the elements
Prior to this commit, the `check` method in the `listOf` type in
`lib/types.nix` would accept *any* list, even if the list was not in
fact a "list of" the specified type.  This meant that for checking
purposes, `listOf bool` and `listOf int` were the same thing.

This commit fixes that problem.