yottadb: nixos-specific fixes
YottaDB does not work correctly on NixOS out of the box
in multiuser environments and requires a special SUID
binary to allow processes to communicate with each other.

This commit is trying to patch YDB to make this happen.

These changes, however, are meant to enable people to use
YottaDB for development on NixOS, but should not be used
in production.