rubberband: fix build on darwin
rubberband's build uses apple's Accelerate framework for it's FFT by
default on darwin. We could either tell it to use the built-in fft (the
code that is used on linux) or add the dependencies on the Apple
frameworks.

This change adds the appropriate dependencies for darwin to build with
rubberband's default configuration.