arm-trusted-firmware: set unfreeIncludeHDCPBlob=false if not used
The `unfreeIncludeHDCPBlob` parameter was introduced as a result of
this reviewer request:

   https://github.com/NixOS/nixpkgs/issues/148890#issuecomment-1032002903

The default value `unfreeIncludeHDCPBlob?true` causes a change in the
`meta.license` field for all of the subpackages within
`pkgs/misc/arm-trusted-firmware/`, and results in them needing
`NIXPKGS_ALLOW_NONFREE=1`.

For Rockchip platforms this change is unavoidable; we are correcting
an incorrect license declaration.  However for non-Rockchip platforms
the file `hdcp.bin` does not get included in the output.  Therefore we
can set `unfreeIncludeHDCPBlob=false` on these platforms and produce
no user-visible change in the resulting output, while preserving the
ability to build them without `NIXPKGS_ALLOW_NONFREE=1`.

Let's do that.