release-lib.nix: fix evaluation with Nix
Previously, release-haskell.nix and release-python.nix could only be
evaluated by Hydra.  But it's useful to be able to evaluate them with
Nix as well, for quick tests of changes to them.

The problem for evaluation with Nix was if a set of derivations (like
haskellPackages) containing a recurseForDerivations attribute was
passed to packagePlatforms, the result would have
recurseForDerivations = {}.  recurseForDerivations being set to a
non-boolean value causes an evaluation failure in Nix (a nasty one
too, without any backtrace).

We can't just remove recurseForDerivations in packagePlatforms,
because its value is important and removing it causes other evaluation
failures, so we have to make sure it's still set to its original value
after.  This also requires a change to _mapTestOnHelper, so that it
ignores the recurseForDerivations attribute of any package set passed
to it, so it doesn't try to find the meta.platforms of the
recurseForDerivations boolean.