matrix-appservice-discord: unquote ${nodejs.src}
This brings it in line with every other instance of nodejs sources being unpacked like this (see `rg 'tar --no-same-owner --no-same-permissions'`). The extracted node-sources is identical in both cases, however due to this difference in the derivation code, two store paths are produced. At a size of roughly 300MB, that's not exactly negligible.