lib/systems: rename examples.nix to all.nix, include alias
When I had to figure out how `lib/systems` worked in order to enable
bootstrapping on mips64el and powerpc64el, by far the most confusing
thing was that nixpkgs' official list of platforms (i.e. those
identifiers which can be used in `pkgsCross.<platform>`) is kept in a
file called `examples.nix`.  When most people, including myself, see a
codebase with files or directories named `examples/` they assume it
contains *examples* which are for tutorial purposes only, and are not
used by the deployed codebase.

This commit attempts to save future nixpkgs contributors the confusion
I encountered.  It renames `lib/systems/examples.nix` to
`lib/systems/all.nix`, since what it contains is *a list of all
systems (or platforms) that nixpkgs currently knows about*.  This
commit also includes an alias in `lib/systems/examples.nix`, so it is
not a breaking change.