lib/source-types.nix: remove redundant sourceType
The provenance type `fromSource` is redundant, and adds numerous
special edge cases.  Let's remove it.  I pointed this out during the
review of #161098; @risicle asked that I submit this as a seperate
subsequent PR in order to not further delay the merge of #161098.
Therefore, I am submitting it now.

The `fromSource` provenance type is redundant; the following are
semantically indistinguishable:

```
  meta = {
    sourceProvenance = [];
  }

  meta = {
    sourceProvenance = [ lib.source-types.fromSource ];
  }
```

Furthermore, `fromSource` gets special treatment: "a package with no
`meta.sourceProvenance` set implies it has no *known* `sourceType`s
**other than `fromSource`.**".

This greatly complicates the "is this package built from source"
routine.  Consider the exemplar `allowNonSourcePredicate` to permit
non-source firmware given in check-meta.nix.  This predicate is likely
to be a very popular policy choice for nixpkgs users, due to the level
of firmware-saturation in consumer-grade hardware.  It changes from:

```
  allowNonSourcePredicate = with lib.lists; pkg: !(any (p: !p.isSource && p!=lib.sourceTypes.binaryFirmware) (toList pkg.meta.sourceProvenance)
```

to

```
  allowNonSourcePredicate = lib.lists.any (p: p!=lib.sourceTypes.binaryFirmware) (toList pkg.meta.sourceProvenance);
```

and, if #175495 merges, even further to:

```
  allowNonSourcePredicate = lib.lists.any (p: p!=lib.sourceTypes.binaryFirmware) pkg.meta.sourceProvenance;
```

Similar simplifications will be enjoyed by any predicates which users
write for themselves, or any "provenance functionals" which express
the `sourceProvenance` of one package as a function of the
`sourceProvenance` of some subpackage.