clang-tools: don't hardcode list of tools
Currently clang-tools' derivation uses a hardcoded list of names to
distinguish between what's a compiler-like binary and what's a tool-like
binary (so that e.g. clang-tidy is included in the derivation, but
clang-13 - not).

Because Clang's tools follow a common naming convention
(clang-somethingsomething), I think it's easier if we just used a simple
regular expression in place of the hardcoded list.