lib/modules: Make extensionOffset handling internal to extendModules
evalModules / extendModules has enough information to perform the
numbering by itself, making extendModules safe and simplifying
submoduleWith.

The `extensionOffset` was supposed to be an implementation detail
of `evalModules`/`extendModules` all along.

Zooming out a bit, the number of `+` operations should match the
number of `extendModules` calls, which can be more than the number
of `submoduleWith` calls, which is less related to the problem being
solved by `extensionOffset`, as it can defer that responsibility
to `extendModules`.