qtscriptgenerator: remove (Qt4)
Currently fetched from google archives, upstream is dead since 2017 and
nothing depends on it.

It has not received any package updates since its introduction in 2012
(as dependency for Amarok), execept for build/infrastructure changes.

amarok-kde4 and thus the dependency was dropped in 2017.

qcad uses qtscriptgenerator internally, but it bundles its own copy.

See #174634.