sweep-visualizer: Remove
`meta.homepage` links to a nonexistent subdomain and
https://scanse.io/downloads still only has instructions for ancient
Ubuntu 14.04.

This package has not seen updates in years and likely never will.

It depends on GNOME2/GTK2 which will go eventually.

It is also marked as unfree, so there's even less incentive to keep
such dead software around.

It crashes on start:
```
$ NIXPKGS_ALLOW_UNFREE=1 nix-shell -p sweep-visualizer
...
$ sweep_visualizer
Segmentation fault (core dumped)
```

See #39976.